backend/requirements.txt
asgiref==3.4.1
autopep8==1.6.0
build==1.2.2.post1
click==8.1.7
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2.13
django-ckeditor==6.3.0
django-ckeditor-5==0.2.13
django-cors-headers==3.11.0
django-crispy-forms==1.13.0
django-import-export==2.5.0
django-jazzmin==2.4.2
django-js-asset==2.2.0
django-mathfilters==1.0.0
django-taggit==1.3.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
et-xmlfile==1.1.0
gunicorn==23.0.0
mysql-connector-python==9.0.0
mysqlclient==2.2.4
odfpy==1.4.1
openpyxl==3.1.5
packaging==24.1
Pillow==9.0.1
pip-tools==7.4.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
PyJWT==2.3.0
PyMySQL==1.1.1
pyproject_hooks==1.2.0
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0.2
setuptools==75.1.0
sqlparse==0.4.2
tablib==3.6.1
toml==0.10.2
typing_extensions==4.12.2
tzdata==2021.5
watchdog==5.0.3
wheel==0.44.0
whitenoise==6.7.0
xlrd==2.0.1
xlwt==1.3.0
